"ads1_m14_a1: 1/4 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ assignments/m14/14.1 Dynamic Median Finding/Solution.java
@@ -7,7 +7,7 @@ class Solution {
 		MinPQ<Double> min = new MinPQ<Double>();
 		MaxPQ<Double> max = new MaxPQ<Double>();
 		double median = 0.0;
-		for (int i = 0; i < number; i++) {
+		for (long i = 0; i < number; i++) {
 			double input = scan.nextDouble();
 			if (input > median) {
 				min.insert(input);
